pyproject.toml
[project]
name = "trenes-arg-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "dotenv>=0.9.9",
    "httpx>=0.28.1",
    "python-telegram-bot[job-queue]>=22.3",
    "supabase>=2.17.0",
]
